requirements.txt
# ============================================================================
# DetectVoice - Requirements File
# Audio Deepfake Detection Suite - All Dependencies
# ============================================================================

# Core Deep Learning
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0

# Transformers and Pre-trained Models
transformers>=4.30.0
datasets>=2.12.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
scipy>=1.10.0
resampy>=0.4.2

# Classical Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
numpy>=1.24.0
pandas>=2.0.0

# Data Augmentation
audiomentations>=0.30.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dimensionality Reduction
umap-learn>=0.5.3
scikit-learn>=1.3.0  # For t-SNE

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Progress & Logging
tqdm>=4.65.0
loguru>=0.7.0

# Experiment Tracking
tensorboard>=2.13.0
mlflow>=2.5.0

# Model Export
onnx>=1.14.0
onnxruntime>=1.15.0
onnxruntime-gpu>=1.15.0  # Optional: GPU support

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0
ipywidgets>=8.0.0

# Web Interface (Optional)
gradio>=3.35.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
pillow>=10.0.0
requests>=2.31.0
wget>=3.2
gdown>=4.7.0
pydub>=0.25.0

# Type Checking
typing-extensions>=4.7.0

# Video Processing (Optional - for multimodal)
opencv-python>=4.8.0

# Advanced Audio Analysis (Optional)
praat-parselmouth>=0.4.3  # For formant analysis
pyworld>=0.3.2  # For pitch analysis

# Distributed Training (Optional)
accelerate>=0.20.0
deepspeed>=0.9.0  # Optional: For large models

# Model Optimization (Optional)
onnx-simplifier>=0.4.0
tensorrt>=8.6.0  # Optional: NVIDIA only

# Huawei MindSpore (Optional)
mindspore-cpu>=2.0.0


# ============================================================================
# Installation Instructions
# ============================================================================
#
# Basic installation:
#   pip install -r requirements.txt
#
# With GPU support:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#   pip install -r requirements.txt
#
# For development:
#   pip install -e .
#   pip install -r requirements-dev.txt
#
# ============================================================================
